refact(news-app): create headlines component
